chore: split `OrderProgressBarV2` unit into smaller units (#5582)
* chore: split `OrderProgressBarV2` unit into smaller units

* chore: break down styled components and extract `DebugPanel`

* chore: breakdown `RenderProgressTopSection` component

* fix: convert default exports to named exports

* fix: import

* fix: import

* fix: address CR comments

* fix: missing text on `finished` step

* fix: expired order background color

* fix: address CR comments

* chore: rename `OrderProgressBarV2` to `OrderProgressBar`

* chore: rename `OrderProgressBarV2Props` to `OrderProgressBarProps`

* chore: rename `useOrderProgressBarV2Props` to `useOrderProgressBarProps`

* chore: rename `ProgressBarV2ExecutingOrdersUpdater` to `ProgressBarExecutingOrdersUpdater`

* fix: refactor functions into Records

* fix: linting

* chore: address CR comments
