--- apps/cowswap-frontend/src/common/hooks/orderProgressBarV2/index.ts
@@ -1,3 +1,0 @@
-export * from './useOrderProgressBarV2Props'
-export * from './types'
-export * from './ProgressBarV2ExecutingOrdersUpdater'
