+++ apps/cowswap-frontend/src/common/pure/OrderProgressBar/steps/AnimatedToken.tsx
@@ -1,0 +1,47 @@
+import { ReactElement } from 'react'
+
+import { TokenWithLogo } from '@cowprotocol/common-const'
+import { TokenLogo } from '@cowprotocol/tokens'
+import { Color, ProductLogo, ProductVariant } from '@cowprotocol/ui'
+import { Currency } from '@uniswap/sdk-core'
+
+import * as styledEl from './styled'
+
+const ICON_SIZE = 136
+const MOBILE_ICON_SIZE = 72
+
+function AnimatedTokens({
+  sellToken,
+  buyToken,
+}: {
+  sellToken: Currency | TokenWithLogo | null | undefined
+  buyToken: Currency | TokenWithLogo | null | undefined
+}): ReactElement {
+  return (
+    <styledEl.AnimatedTokensWrapper>
+      <styledEl.TokenWrapper position="left" size={ICON_SIZE} sizeMobile={MOBILE_ICON_SIZE}>
+        <TokenLogo token={sellToken} size={ICON_SIZE} sizeMobile={MOBILE_ICON_SIZE} />
+      </styledEl.TokenWrapper>
+      <styledEl.TokenWrapper position="center" size={ICON_SIZE} sizeMobile={MOBILE_ICON_SIZE}>
+        <TokenLogo token={buyToken} size={ICON_SIZE} sizeMobile={MOBILE_ICON_SIZE} />
+      </styledEl.TokenWrapper>
+      <styledEl.TokenWrapper
+        position="right"
+        bgColor={Color.cowfi_blue_dark}
+        size={ICON_SIZE}
+        sizeMobile={MOBILE_ICON_SIZE}
+      >
+        <ProductLogo
+          variant={ProductVariant.CowSwap}
+          theme={'dark'}
+          overrideHoverColor={'#65D9FF'}
+          height={58}
+          heightMobile={32}
+          logoIconOnly
+        />
+      </styledEl.TokenWrapper>
+    </styledEl.AnimatedTokensWrapper>
+  )
+}
+
+export default AnimatedTokens
