+++ apps/cowswap-frontend/src/common/pure/OrderProgressBar/steps/OrderIntent.tsx
@@ -1,0 +1,53 @@
+import { isSellOrder } from '@cowprotocol/common-utils'
+import { TokenLogo } from '@cowprotocol/tokens'
+import { TokenAmount } from '@cowprotocol/ui'
+import { CurrencyAmount } from '@uniswap/sdk-core'
+
+import { Order } from 'legacy/state/orders/actions'
+
+import * as styledEl from './styled'
+
+const TOKEN_SIZE = 20
+
+export function OrderIntent({ order }: { order?: Order }) {
+  if (!order) return null
+
+  const { inputToken, outputToken, kind, sellAmount, buyAmount } = order
+
+  if (!inputToken || !outputToken || !sellAmount || !buyAmount) {
+    return null
+  }
+
+  const isSell = isSellOrder(kind)
+
+  const sellCurrencyAmount = CurrencyAmount.fromRawAmount(inputToken, sellAmount)
+  const buyCurrencyAmount = CurrencyAmount.fromRawAmount(outputToken, buyAmount)
+
+  const sellTokenPart = (
+    <>
+      <TokenLogo token={inputToken} size={TOKEN_SIZE} />
+      <TokenAmount amount={sellCurrencyAmount} tokenSymbol={inputToken} />
+    </>
+  )
+
+  const buyTokenPart = (
+    <>
+      <TokenLogo token={outputToken} size={TOKEN_SIZE} />
+      <TokenAmount amount={buyCurrencyAmount} tokenSymbol={outputToken} />
+    </>
+  )
+
+  return (
+    <styledEl.OriginalOrderIntent>
+      {isSell ? (
+        <>
+          {sellTokenPart} for at least {buyTokenPart}
+        </>
+      ) : (
+        <>
+          {buyTokenPart} for at most {sellTokenPart}
+        </>
+      )}
+    </styledEl.OriginalOrderIntent>
+  )
+}
