+++ apps/cowswap-frontend/src/common/pure/OrderProgressBar/steps/styled.ts
@@ -1,12 +1,15 @@
 import IMAGE_STAR_SHINE from '@cowprotocol/assets/cow-swap/star-shine.svg'
 import { SingleLetterLogoWrapper } from '@cowprotocol/tokens'
-import { ButtonPrimary, Font, LinkStyledButton, Media, UI } from '@cowprotocol/ui'
+import { ButtonPrimary, Color, Font, LinkStyledButton, Media, UI } from '@cowprotocol/ui'
 
 import styled, { css, keyframes } from 'styled-components/macro'
 
-import { CancelButton as CancelButtonOriginal } from '../CancelButton'
+export { Description, ProgressImageWrapper } from '../sharedStyled'
 
-const BLUE_COLOR = '#65d9ff'
+import { CancelButton as CancelButtonOriginal } from '../../CancelButton'
+import { ProgressImageWrapper } from '../sharedStyled'
+
+const BLUE_COLOR = Color.cowfi_blue_lighter
 
 const progressAnimation = (startAt: number, end: number) => {
   const start = end - startAt
@@ -21,32 +24,6 @@ const progressAnimation = (startAt: number, end: number) => {
   `
 }
 
-const sweatDropAnimation = keyframes`
-  0% {
-    transform: translateY(0);
-    opacity: 1;
-  }
-  80% {
-    transform: translateY(20px);
-    opacity: 1;
-  }
-  100% {
-    transform: translateY(40px);
-    opacity: 0;
-  }
-`
-
-const getOpacity = (status: string, isDarkMode: boolean): number => {
-  const opacityMap = {
-    done: isDarkMode ? 0.3 : 0.1,
-    active: 1,
-    next: isDarkMode ? 0.6 : 0.5,
-    future: isDarkMode ? 0.3 : 0.2,
-    disabled: 0.2,
-  }
-  return opacityMap[status as keyof typeof opacityMap] || 1
-}
-
 export const ProgressContainer = styled.div`
   width: 100%;
   height: auto;
@@ -54,70 +31,6 @@ export const ProgressContainer = styled.div`
   flex-direction: column;
   align-items: flex-start;
   padding: 0 0 24px;
-`
-
-export const StepsContainer = styled.div<{ $height: number; $minHeight?: string; bottomGradient?: boolean }>`
-  position: relative;
-  height: ${({ $height }) => $height}px;
-  min-height: ${({ $minHeight }) => $minHeight || '192px'};
-  overflow: hidden;
-  transition: height 0.5s ease-in-out;
-  width: 100%;
-  padding: 0;
-
-  // implement a gradient to hide the bottom of the steps container using white to opacity white using pseudo element
-  &::after {
-    content: ${({ bottomGradient }) => (bottomGradient ? '""' : 'none')};
-    position: absolute;
-    bottom: 0;
-    left: 0;
-    width: 100%;
-    height: 30px;
-    background: linear-gradient(to bottom, transparent, var(${UI.COLOR_PAPER}));
-  }
-`
-
-export const StepsWrapper = styled.div`
-  display: flex;
-  flex-flow: column nowrap;
-  padding: 0;
-  width: 100%;
-  position: relative;
-  transition: transform 1s ease-in-out;
-`
-
-export const Step = styled.div<{ status: string; isFirst: boolean }>`
-  transition: opacity 0.3s ease-in-out;
-  display: flex;
-  align-items: flex-start;
-  margin: 0 auto;
-  width: 100%;
-  padding: 30px 30px 10px;
-  opacity: ${({ status, theme }) => getOpacity(status, theme.darkMode)};
-`
-
-export const Content = styled.div`
-  display: flex;
-  flex-direction: column;
-`
-
-export const Title = styled.h3<{ customColor?: string }>`
-  color: ${({ customColor }) => customColor || `var(${UI.COLOR_TEXT_PAPER})`};
-  margin: 0;
-  font-size: 21px;
-`
-
-export const Description = styled.div<{ center?: boolean; margin?: string }>`
-  margin: ${({ margin }) => margin || '8px 0 0'};
-  font-size: 14px;
-  color: var(${UI.COLOR_TEXT_OPACITY_70});
-  text-align: ${({ center }) => (center ? 'center' : 'left')};
-
-  > button,
-  > a {
-    text-decoration: underline;
-    color: var(${UI.COLOR_TEXT_PAPER});
-  }
 `
 
 export const Link = styled.a<{ underline?: boolean }>`
@@ -152,71 +65,6 @@ export const Button = styled(LinkStyledButton)`
   }
 `
 
-export const ProgressImageWrapper = styled.div<{ bgColor?: string; padding?: string; height?: string; gap?: string }>`
-  width: 100%;
-  height: ${({ height }) => height || '246px'};
-  min-height: 200px;
-  max-height: ${({ height }) => height || '246px'};
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-flow: column wrap;
-  border-radius: 21px;
-  padding: ${({ padding }) => padding || '0'};
-  gap: ${({ gap }) => gap || '0'};
-  background: ${({ bgColor }) => bgColor || `var(${UI.COLOR_PAPER_DARKER})`};
-  transition: height 0.3s ease-in-out;
-  position: relative;
-  overflow: hidden;
-
-  ${Media.upToSmall()} {
-    min-height: auto;
-    height: auto;
-  }
-
-  > img,
-  > svg {
-    --size: 100%;
-    max-width: var(--size);
-    max-height: var(--size);
-    height: var(--size);
-    width: var(--size);
-    object-fit: contain;
-    padding: 0;
-    margin: 0;
-  }
-
-  > div {
-    display: flex;
-  }
-`
-
-export const DebugPanel = styled.div`
-  position: fixed;
-  bottom: 10px;
-  right: 10px;
-  background-color: rgba(0, 0, 0, 0.7);
-  color: white;
-  padding: 10px;
-  border-radius: 5px;
-  z-index: 1000;
-`
-
-export const ProgressTopSection = styled.div`
-  position: relative;
-  width: 100%;
-  display: flex;
-  flex-flow: column wrap;
-  align-items: center;
-  border-radius: 21px;
-  background: var(${UI.COLOR_PAPER_DARKER});
-  min-height: 230px;
-
-  ${Media.upToSmall()} {
-    min-height: auto;
-  }
-`
-
 export const OriginalOrderIntent = styled.span`
   display: flex;
   flex-flow: row wrap;
@@ -233,14 +81,6 @@ export const OriginalOrderIntent = styled.span`
     background: ${({ theme }) => (theme.darkMode ? `var(${UI.COLOR_TEXT})` : `var(${UI.COLOR_PAPER_DARKEST})`)};
     color: ${({ theme }) => (theme.darkMode ? `var(${UI.COLOR_PAPER_DARKER})` : `var(${UI.COLOR_TEXT})`)};
   }
-`
-
-export const OrderTokenImage = styled.img`
-  --size: 20px;
-  width: var(--size);
-  height: var(--size);
-  border-radius: var(--size);
-  background: var(${UI.COLOR_PAPER_DARKEST});
 `
 
 export const AnimatedTokensWrapper = styled.div`
@@ -271,10 +111,11 @@ export const TokenWrapper = styled.div<{
   right: 0;
   left: 0;
   margin: auto;
+  will-change: transform;
   animation: ${({ position }) => (position === 'left' ? 'appear-left' : position === 'right' ? 'appear-right' : 'none')}
     2.5s cubic-bezier(0.19, 1, 0.22, 1) forwards;
   animation-delay: ${({ position }) => (position === 'center' ? '0s' : '0.75s')};
-  border: ${({ position }) => (position === 'right' || 'center' ? `8px solid ${BLUE_COLOR}` : '0')};
+  border: ${({ position }) => (position === 'right' || position === 'center' ? `8px solid ${BLUE_COLOR}` : '0')};
   box-sizing: content-box;
   background: ${({ position, bgColor }) =>
     position === 'right' ? bgColor || `var(${UI.COLOR_PRIMARY})` : 'transparent'};
@@ -383,104 +224,6 @@ export const ConclusionContent = styled.div`
   padding: 0;
   width: 100%;
   margin: 20px auto 0;
-`
-
-export const CowImage = styled.div`
-  height: 100%;
-  width: auto;
-  display: flex;
-  align-items: flex-end;
-  justify-content: flex-start;
-  position: relative;
-
-  ${Media.upToSmall()} {
-    width: 100%;
-    align-items: center;
-    justify-content: center;
-    max-height: 100%;
-  }
-
-  > svg {
-    height: 100%;
-    width: 100%;
-    max-width: 199px;
-
-    ${Media.upToSmall()} {
-      max-width: 100%;
-    }
-  }
-`
-
-export const TokenPairTitle = styled.span`
-  margin: 4px 0 4px 4px;
-  background: #99ecff;
-  border-radius: 12px;
-  padding: 0 6px;
-  word-break: break-word;
-  line-height: 1;
-`
-
-export const TokenImages = styled.div`
-  position: absolute;
-  bottom: 10px;
-  right: 10px;
-  display: flex;
-  gap: 5px;
-
-  > div {
-    border: 3px solid #65d9ff;
-    box-sizing: content-box;
-  }
-
-  > div:first-child {
-    margin-right: -20px;
-  }
-`
-
-export const Surplus = styled.div`
-  font-weight: bold;
-  color: inherit;
-  width: 100%;
-  height: 100%;
-  text-align: right;
-  line-height: 1.2;
-  margin: 0;
-`
-
-export const FinishedTagLine = styled.div`
-  line-height: 1.2;
-  font-weight: bold;
-  color: inherit;
-  max-width: 100%;
-  font-size: 22px;
-  width: 100%;
-  text-align: right;
-  height: 100%;
-  display: flex;
-  align-items: center;
-  overflow: hidden;
-  padding: 0;
-  flex: 1 1 0;
-
-  ${Media.upToSmall()} {
-    flex: 0 0 auto;
-    height: auto;
-  }
-`
-
-export const FinishedLogo = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  z-index: 1;
-  padding: 0;
-  margin: auto 0 0;
-  width: 100%;
-  flex: 0;
-
-  > b {
-    font-weight: 700;
-  }
 `
 
 export const ShareButton = styled(ButtonPrimary)`
@@ -677,7 +420,7 @@ export const ReceivedAmount = styled.span`
   }
 `
 
-export const SoldAmount = styled(ReceivedAmount)``
+export const SoldAmount = ReceivedAmount
 
 export const ExtraAmount = styled.p`
   display: flex;
@@ -769,154 +512,6 @@ export const CancellationFailedBanner = styled.div`
   font-size: 15px;
 `
 
-export const NumberedElement = styled.div<{
-  status: string
-  customColor?: string
-  $isUnfillable?: boolean
-  $isCancelling?: boolean
-}>`
-  --size: 28px;
-  width: var(--size);
-  height: var(--size);
-  min-width: var(--size);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin-right: 15px;
-  color: ${({ status }) => (status === 'active' ? `var(${UI.COLOR_PAPER})` : `var(${UI.COLOR_PAPER})`)};
-  font-weight: bold;
-  font-size: 16px;
-  background-color: ${({ status, customColor, $isUnfillable, $isCancelling }) =>
-    $isCancelling
-      ? `var(${UI.COLOR_DANGER_BG})`
-      : $isUnfillable
-        ? '#996815'
-        : customColor || (status === 'active' ? '#2196F3' : `var(${UI.COLOR_TEXT})`)};
-  border-radius: 50%;
-  position: relative;
-`
-
-export const Spinner = styled.div`
-  position: absolute;
-  top: -4px;
-  left: -4px;
-  right: -4px;
-  bottom: -4px;
-  border: 2px solid transparent;
-  border-top-color: ${`var(${UI.COLOR_PRIMARY_LIGHTER})`};
-  border-radius: 50%;
-  animation: spin 1.5s cubic-bezier(0.25, 0.46, 0.45, 0.94) infinite;
-
-  @keyframes spin {
-    0% {
-      transform: rotate(0deg);
-    }
-    100% {
-      transform: rotate(360deg);
-    }
-  }
-`
-
-export const SweatDrop = styled.div`
-  color: #ffffff;
-  position: absolute;
-  left: 160px;
-  top: 75px;
-  width: 15px;
-  height: 22px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  animation: ${sweatDropAnimation} 2s cubic-bezier(0.19, 1, 0.22, 1) infinite;
-
-  > svg {
-    width: 100%;
-    height: 100%;
-    color: inherit;
-  }
-`
-
-export const ClockAnimation = styled.div`
-  --size: 85px;
-  width: var(--size);
-  height: var(--size);
-  position: absolute;
-  bottom: 36px;
-  right: 71px;
-  background: #996815;
-  border-radius: var(--size);
-  padding: 6px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-
-  ${Media.upToSmall()} {
-    bottom: 16px;
-    right: 16px;
-  }
-`
-
-export const FinishedImageContent = styled.div`
-  display: flex;
-  flex-flow: column wrap;
-  align-items: center;
-  justify-content: space-between;
-  height: 100%;
-  width: 50%;
-  position: relative;
-  border: 2px solid #99ecff;
-  border-radius: 21px;
-  padding: 12px;
-  gap: 14px;
-  color: ${({ theme }) => (theme.darkMode ? `var(${UI.COLOR_BUTTON_TEXT})` : `var(${UI.COLOR_TEXT})`)};
-
-  ${Media.upToSmall()} {
-    width: 100%;
-    max-height: 290px;
-  }
-`
-
-export const BenefitSurplusContainer = styled.div`
-  width: 100%;
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: flex-end;
-  text-align: right;
-  justify-content: flex-start;
-  gap: 0;
-  font-size: 20px;
-  line-height: 1.4;
-`
-
-export const BenefitText = styled.div`
-  width: 100%;
-  height: 100%;
-  overflow: hidden;
-  padding: 6px 0 0;
-  margin: 0;
-  box-sizing: border-box;
-
-  > span {
-    display: inline-block;
-    line-height: 1.2;
-    text-align: left;
-    word-break: break-word;
-    hyphens: auto;
-    width: 100%;
-  }
-`
-
-export const BenefitTagLine = styled.div`
-  width: auto;
-  font-size: 14px;
-  margin: 0 auto auto 0;
-  border-radius: 12px;
-  padding: 2px 10px;
-  background-color: #3fc4ff;
-  color: #000000;
-`
-
 export const CircularProgress = styled.svg`
   width: 100%;
   height: 100%;
@@ -930,6 +525,7 @@ export const CircleProgress = styled.circle<{ startAt: number; end: number }>`
   stroke: #05a1ff;
   stroke-width: 6;
   stroke-linecap: round;
+  stroke-dasharray: 283; /* Approximately 2 * PI * 45 */
 
   ${({ startAt, end }) => css`
     animation: ${progressAnimation(startAt, end)} ${startAt}s linear infinite;
